Creating a frameWork

+++ src/test/java/base/Global.java
@@ -2,14 +2,23 @@ package base;
 
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
+import org.testng.annotations.AfterSuite;
+import org.testng.annotations.AfterTest;
 
 public class Global {
-    WebDriver driver;
+    static WebDriver driver;
     public WebDriver getDriver(){
-        driver = new ChromeDriver();
-        driver.manage().window().maximize();
-        driver.manage().deleteAllCookies();
-        driver.get("https://rahulshettyacademy.com/client");
+        if(driver==null){
+            driver = new ChromeDriver();
+            driver.manage().window().maximize();
+            driver.manage().deleteAllCookies();
+            driver.get("https://rahulshettyacademy.com/client");
+            return driver;
+        }
         return driver;
+    }
+    @AfterSuite
+    void terDown(){
+        driver.quit();
     }
 }
